Reapply "19일 RESTful 계속, Controller 와 front 사이에 정보가 어떻게 오고가는지 curd 하면서 연습 read.html 에 4~6번 작업하면 끝"
This reverts commit b187671862ea8d953a2804b6ccec01e787cd666c.

+++ SpringBasic/bootPractice/src/main/java/com/busanit501/bootpractice/repository/ReplyRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface ReplyRepository extends JpaRepository<Reply, Integer> {
+public interface ReplyRepository extends JpaRepository<Reply, Long> {
 
     @Query("select r from Reply r where r.board.bno = :bno")
     Page<Reply> listOfBoard(Long bno, Pageable pageable);
